+++ src/graphTheory/component_array.js
@@ -9,16 +9,16 @@ class ComponentArray extends RobustArray.SetifyType(NodeArray) {
      * adds a component to the graph if not present
      * @param {Component} compArg [description]
      */
-    addComponent(compArg) {
-        (!this.contains(compArg) && this.hasIntersectingComponent(compArg)) ? this.integrateComponent(compArg) : this.components.push(compArg);
+    push(compArg) {
+        this.hasIntersectingComponent(compArg) ? this.integrateComponent(compArg) : super.push(compArg);
     }
-    // removeComponent(compArg) {
-    // let cPos = this.components.findIndex(currComp => currComp.isEquivalent(compArg));
-    // return cPos > -1 ? this.components.splice(cPos, 1) : compArg;
-    // }
-    removeElement(argObj) {
-        let eqIdx = this.findIndex(el => el.isEquivalent(argObj));
-        return eqIdx > -1 && this.splice(eqIdx, 1);
+    /**
+     * checks if any current components share nodes with the argument
+     * @param  {Component}  compArg [description]
+     * @return {Boolean}         [description]
+     */
+    hasIntersectingComponent(compArg) {
+        return this.excludeElement(compArg).some(currComp => currComp.intersects(compArg));
     }
     /**
      * returns any current components which intersect with the specified component
@@ -26,7 +26,7 @@ class ComponentArray extends RobustArray.SetifyType(NodeArray) {
      * @return {Component} the first intersecting component
      */
     findIntersectingComponent(compArg) {
-        return this.find(currComp => currComp.intersects(compArg));
+        return this.excludeElement(compArg).find(currComp => currComp.intersects(compArg));
     }
     /**
      * combines the nodes of two intersecting components
@@ -43,18 +43,10 @@ class ComponentArray extends RobustArray.SetifyType(NodeArray) {
      * @param  {Component} compArg [description]
      */
     integrateComponent(compArg) {
-        let oComp = this.findIntersectingComponent(compArg);
-        if (oComp) this.mergeComponents(oComp, compArg);
-        // return oComp && this.mergeComponents(oComp, compArg);
+        let iComp = this.findIntersectingComponent(compArg);
+        return (!!iComp) && this.mergeComponents(iComp, compArg);
     }
-    /**
-     * checks if any current components share nodes with the argument
-     * @param  {Component}  compArg [description]
-     * @return {Boolean}         [description]
-     */
-    hasIntersectingComponent(compArg) {
-        return this.components.some(currComp => currComp.intersects(compArg));
-    }
+
 };
 
 module.exports = ComponentArray;